requirements.txt
requests==2.13.0
beautifulsoup4==4.5.3
PyYAML==3.12
